fix: count 오류 수정

+++ src/main/java/com/hack/stock2u/product/dto/ProductCountProjection.java
@@ -5,7 +5,7 @@ import com.hack.stock2u.constant.ProductType;
 import java.util.Date;
 
 public interface ProductCountProjection {
-  Integer getTotalCount();
+  Long getTotalCount();
 
   Double getDistance();
 }
